atomic-docker/project/functions/python_api_service/linkedin_service.py
import os
import logging
from typing import Optional, List, Dict, Any

# Configure logging for this module
logger = logging.getLogger(__name__)

# --- LinkedIn Package Compatibility Layer ---
# Handle different import path variations for python-linkedin
LinkedInApplication = None
LinkedInAuthentication = None
LinkedInError = None

try:
    # Try primary import path
    from linkedin.linkedin import (
        LinkedInApplication as LinkedInApp,
        LinkedInAuthentication as LinkedInAuth,
        LinkedInError as LinkedInErr,
    )
    LinkedInApplication = LinkedInApp
    LinkedInAuthentication = LinkedInAuth
    LinkedInError = LinkedInErr
    logger.info("Imported python-linkedin from linkedin.linkedin")
except ImportError:
    try:
        # Alternative import path
        from linkedin import (
            LinkedInApplication as LinkedInApp,
            LinkedInAuthentication as LinkedInAuth,
            LinkedInError as LinkedInErr,
        )
        LinkedInApplication = LinkedInApp
        LinkedInAuthentication = LinkedInAuth
        LinkedInError = LinkedInErr
        logger.info("Imported python-linkedin from linkedin")
    except ImportError:
        logger.warning("python-linkedin library not found. Using mock implementation.")

        # Create mock classes for development/testing without actual dependencies
        class LinkedInError(Exception):
            """Mock LinkedIn error class."""
            pass

        class LinkedInAuthentication:
            """Mock authentication handler for testing."""
            def __init__(self, consumer_key: str, consumer_secret: str,
                         user_token: str, user_secret: str, return_url: str):
                self.consumer_key = consumer_key
                self.consumer_secret = consumer_secret
                self.user_token = user_token
                self.user_secret = user_secret
                self.return_url = return_url

        class LinkedInApplication:
            """Mock LinkedIn application client."""
            def __init__(self, authentication):
                self.auth = authentication

            def get_profile(self, member_id: Optional[str] = None,
                          selectors: Optional[List[str]] = None) -> Dict[str, Any]:
                """Mock profile retrieval."""
                logger.debug(f"Mock get_profile called for member_id: {member_id}")
                return {
                    'id': member_id or 'mock-user-id',
                    'firstName': 'Mock',
                    'lastName': 'User',
                    'headline': 'Software Engineer at Mock Company',
                    'publicProfileUrl': 'https://www.linkedin.com/in/mockuser',
                }

            def submit_share(self, comment: str, title: str,
                           description: str, submitted_url: str) -> Dict[str, str]:
                """Mock share submission."""
                logger.debug(f"Mock submit_share called with title: {title}")
                return {
                    'updateKey': 'mock-update-key-12345',
                    'updateUrl': 'https://www.linkedin.com/updates/mock-update-12345',
                }

# Make classes available at module level
if LinkedInApplication is None:
    LinkedInApplication = LinkedInApp = LinkedInApplication
if LinkedInError is None:
    LinkedInError = LinkedInErr = LinkedInError
if LinkedInAuthentication is None:
    LinkedInAuthentication = LinkedInAuth = LinkedInAuthentication

# --- Service Implementation ---

async def get_linkedin_api(user_id: str, db_conn_pool=None) -> Optional[LinkedInApplication]:
    """
    Constructs and returns a LinkedIn API client for the given user.

    Args:
        user_id: The user identifier
        db_conn_pool: Database connection pool (unused in mock mode)

    Returns:
        LinkedInApplication instance or None if credentials are missing
    """
    # Get credentials from environment
    consumer_key = os.environ.get("LINKEDIN_CONSUMER_KEY", "").strip()
    consumer_secret = os.environ.get("LINKEDIN_CONSUMER_SECRET", "").strip()
    user_token = os.environ.get("LINKEDIN_USER_TOKEN", "").strip()
    user_secret = os.environ.get("LINKEDIN_USER_SECRET", "").strip()

    if not all([consumer_key, consumer_secret, user_token, user_secret]):
        logger.error("LinkedIn API credentials incomplete")
        return None

    try:
        auth = LinkedInAuthentication(
            consumer_key=consumer_key,
            consumer_secret=consumer_secret,
            user_token=user_token,
            user_secret=user_secret,
            return_url='urn:ietf:wg:oauth:2.0:oob'
        )
        return LinkedInApplication(auth)

    except LinkedInError as e:
        logger.error(f"LinkedIn initialization error for user {user_id}: {e}")
        return None
    except Exception as e:
        logger.error(f"Unexpected error initializing LinkedIn client: {e}")
        return None

async def search_linkedin_profiles(api, query: str) -> List[Dict[str, Any]]:
    """
    Placeholder for LinkedIn profile search.

    Note: python-linkedin is outdated for LinkedIn People Search API v2.
    This function returns empty results and logs the limitation.
    """
    logger.warning("LinkedIn profile search not implemented (library outdated")
    return []

# For backwards compatibility
get_linkedin_api_sync = get_linkedin_api
