atomic-docker/project/functions/python_api_service/requirements.txt
# Python 3.8+ compatible requirements
tweepy>=4.14.0,<5.0.0
google-api-python-client>=2.0.0,<3.0.0
google-auth-httplib2>=0.1.0,<1.0.0
google-auth-oauthlib>=1.0.0,<2.0.0
cryptography>=41.0.0,<43.0.0
openai>=1.0.0,<2.0.0
requests>=2.28.0,<3.0.0
python-linkedin>=1.0.0,<2.0.0
pandas>=1.5.0,<3.0.0
flask>=2.3.0,<3.0.0
flask-sock>=0.6.0,<1.0.0
pvporcupine>=3.0.0,<3.1.0
websockets>=11.0,<12.0
lancedb>=0.5.3,<1.0.0
python-dotenv>=1.0.0,<2.0.0
uuid>=1.30
notion-client>=2.0.0,<3.0.0
psycopg2-binary>=2.9.0,<3.0.0
beautifulsoup4>=4.12.0,<5.0.0
lxml>=4.9.0,<6.0.0
simple-salesforce>=1.12.0,<2.0.0
xero-python>=1.5.0,<2.0.0
pyowm>=3.3.0,<4.0.0
newsapi-python>=0.2.6,<1.0.0
quickbooks-python>=1.0.0,<2.0.0
mailchimp-marketing>=3.0.0,<4.0.0
zenpy>=2.0.0,<3.0.0
greenhouse-python>=1.0.0,<2.0.0
docusign-esign>=3.0.0,<4.0.0
jira>=3.5.0,<4.0.0
PyGithub>=1.59.0,<2.0.0
pybamboohr>=0.8.0,<1.0.0
python-wordpress-xmlrpc>=2.6.0,<3.0.0
kafka-python>=2.0.0,<3.0.0
elevenlabs>=0.2.0,<1.0.0
